+++ app/src/main/kotlin/com/prajwalch/torrentsearch/ui/home/component/VoiceInputButton.kt
@@ -3,9 +3,9 @@ package com.prajwalch.torrentsearch.ui.home.component
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Icon
+import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -16,23 +16,18 @@ import com.prajwalch.torrentsearch.R
 import com.prajwalch.torrentsearch.ui.tv.tvFocusHighlight
 
 @Composable
-fun SearchButton(
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-    enabled: Boolean = true,
-) {
-    Button(
-        modifier = modifier.tvFocusHighlight(enabled = enabled),
+fun VoiceInputButton(onClick: () -> Unit, modifier: Modifier = Modifier) {
+    OutlinedButton(
+        modifier = modifier.tvFocusHighlight(),
         onClick = onClick,
-        enabled = enabled,
         contentPadding = ButtonDefaults.ButtonWithIconContentPadding,
     ) {
         Icon(
             modifier = Modifier.size(ButtonDefaults.IconSize),
-            painter = painterResource(R.drawable.ic_search),
+            painter = painterResource(R.drawable.ic_mic),
             contentDescription = null,
         )
         Spacer(modifier = Modifier.width(ButtonDefaults.IconSpacing))
-        Text(text = stringResource(R.string.home_button_search))
+        Text(stringResource(R.string.home_button_voice_input))
     }
 }